extract SolutionFileBuilder
